requirements.txt
streamlit==1.35.0
streamlit-webrtc==0.47.1
opencv-python-headless
mediapipe==0.10.9
av==10.0.0
numpy<2
protobuf==3.20.3
tensorflow-cpu==2.12.0
